refactor(backend): extract direct document preview rebind
- Move direct-node document preview host-action rebinding into a focused module.\n- Keep compatibility exports in direct_node_runtime.py for private callers/tests.\n- Preserve preview-only forced host-action behavior.
